Vault Integration
When env variable 'secret_store' is on, the credentials are read from Vault.

Fix: https://github.com/edgexfoundry/docker-edgex-mongo/issues/34

Note1:
- Vault should be unsealed.
- '/vault/config/assets/resp-init.json' file should be present (contains vault master token)
Because of the two above, changes need to be done in the compose-file.
Edgex-mongo container should depend on 'vault-worker' and use 'vault-config' volume.

Note2:
- The security config toml file contains 'Server = "edgex-vault"', that works for docker,
but probably will not work for snap.
I will wait for https://github.com/edgexfoundry/docker-edgex-mongo/pull/42, in order to
decide how to handle it.

Because of still missing functionality for populating the database with the credentials,
the default value of the 'secret-store' variable is Off.
To be able to test it in the dockerized environment, I have created a contaner that
populates the database with needed credentials (useing vault-hashicorp client)(not committed anywhere)

Signed-off-by: difince <dianaa@vmware.com>