Add --configDir, --profile cmd arguments
1. security config file - removed
Now available two config files - for docker or snap -  with the all same sections.
The only difference is what hostname is used for vault - localhost (snap,dev), edgex-vault (docker)

2. Changes dano in cmd arguments
--config cmd argument does not exists anymore. (affect the current implementation of snap)
The client is supposed to provide --configDir direct or relative path to config directory
If configDir is not provided relative 'res' dir path is used.

if --profile is avaibale it is added to the configDir path

At the end, the 'configuration.toml' file name is added.
ConfigFileLocation=configDir/profile?/configuration.toml

Signed-off-by: difince <dianaa@vmware.com>